Move ActivationConflictException to config-provisioning
Move, so that it can be used in MaintenanceDeployment
