+++ config-provisioning/src/main/java/com/yahoo/config/provision/exception/ActivationConflictException.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.config.provision.exception;
 
 /**
  * Exception used when activation cannot be done because activation is for
@@ -9,7 +9,9 @@ package com.yahoo.vespa.config.server;
  * @author hmusum
  */
 public class ActivationConflictException extends RuntimeException {
-    public ActivationConflictException(String s) {
-        super(s);
-    }
+
+    public ActivationConflictException(String message) { super(message); }
+
+    public ActivationConflictException(String message, Throwable cause) { super(message, cause); }
+
 }
